プッシュ通知関連の処理を別ファイルに切り出し
