+++ example-app/SantokuApp/src/components/initialize/helpers/index.ts
@@ -3,3 +3,4 @@ export * from './initialData';
 export * from './navigation';
 export * from './bundledMessage';
 export * from './splashScreen';
+export * from './notification';
